+++ CHANGELOG.md
@@ -11,6 +11,9 @@
 
 ### Bug fixes
 
+`@algorand-builder/runtime`
+    * Remove asset holding from account if `closeRemainderTo` is specified.
+    * Asset creator should not be able to close it's holding to another account.
 
 ## v1.0.2 2021-05-18